@@ -3,5 +3,5 @@ package proxyd
 import "fmt"
 
 func wrapErr(err error, msg string) error {
-	return fmt.Errorf("%s %v", msg, err)
+	return fmt.Errorf("%s %w", msg, err)
 }
